+++ source/src/vm/pc6001/sub.cpp
@@ -380,7 +380,7 @@ void SUB::save_state(FILEIO* state_fio)
 		state_fio->FputInt32(length_tmp);
 		while(length_tmp != 0) {
 			uint8 buffer_tmp[1024];
-			int length_rw = min(length_tmp, sizeof(buffer_tmp));
+			int length_rw = min(length_tmp, (int)sizeof(buffer_tmp));
 			fio->Fread(buffer_tmp, length_rw, 1);
 			state_fio->Fwrite(buffer_tmp, length_rw, 1);
 			length_tmp -= length_rw;
@@ -419,7 +419,7 @@ bool SUB::load_state(FILEIO* state_fio)
 		fio->Fopen(rec_file_path, FILEIO_READ_WRITE_NEW_BINARY);
 		while(length_tmp != 0) {
 			uint8 buffer_tmp[1024];
-			int length_rw = min(length_tmp, sizeof(buffer_tmp));
+			int length_rw = min(length_tmp, (int)sizeof(buffer_tmp));
 			state_fio->Fread(buffer_tmp, length_rw, 1);
 			if(fio->IsOpened()) {
 				fio->Fwrite(buffer_tmp, length_rw, 1);
